--- a/TurnerSoftware.RobotsExclusionTools.sln
+++ b/TurnerSoftware.RobotsExclusionTools.sln
@@ -16,7 +16,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TurnerSoftware.RobotsExclus
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "global", "global", "{969CA4D9-8FB0-4D39-8FEA-ADC663F15205}"
 	ProjectSection(SolutionItems) = preProject
-		.appveyor.yml = .appveyor.yml
 		.codecov.yml = .codecov.yml
 		.editorconfig = .editorconfig
 		.gitignore = .gitignore