--- a/src/Directory.Build.props
+++ b/src/Directory.Build.props
@@ -22,7 +22,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
-    <PackageReference Include="TurnerSoftware.BuildVersioning" Version="0.3.0" PrivateAssets="All" />
+    <PackageReference Include="TurnerSoftware.BuildVersioning" Version="0.4.0" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup>