--- a/src/TurnerSoftware.RobotsExclusionTools/AssemblyInternals.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/AssemblyInternals.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("TurnerSoftware.RobotsExclusionTools.Tests")]