--- a/src/TurnerSoftware.RobotsExclusionTools/Constants.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Constants.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace TurnerSoftware.RobotsExclusionTools;
+
+public static class Constants
+{
+	public const string UserAgentField = "User-agent";
+	public const string DisallowField = "Disallow";
+	public const string AllowField = "Allow";
+
+	public const string CrawlDelayField = "Crawl-delay";
+	public const string SitemapField = "Sitemap";
+
+	public const string UserAgentWildcard = "*";
+}