--- a/src/TurnerSoftware.RobotsExclusionTools/Constants.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Constants.cs
@@ -5,6 +5,7 @@ namespace TurnerSoftware.RobotsExclusionTools;
 public static class Constants
 {
 	public const string UserAgentField = "User-agent";
+
 	public const string DisallowField = "Disallow";
 	public const string AllowField = "Allow";
 
@@ -12,4 +13,15 @@ public static class Constants
 	public const string SitemapField = "Sitemap";
 
 	public const string UserAgentWildcard = "*";
+
+	public static class UTF8
+	{
+		public static ReadOnlySpan<byte> UserAgentField => "User-agent"u8;
+
+		public static ReadOnlySpan<byte> DisallowField => "Disallow"u8;
+		public static ReadOnlySpan<byte> AllowField => "Allow"u8;
+
+		public static ReadOnlySpan<byte> CrawlDelayField => "Crawl-delay"u8;
+		public static ReadOnlySpan<byte> SitemapField => "Sitemap"u8;
+	}
 }