--- a/src/TurnerSoftware.RobotsExclusionTools/Helpers/RobotsExclusionProtocolHelper.cs
+++ b/src/TurnerSoftware.RobotsExclusionTools/Helpers/RobotsExclusionProtocolHelper.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Text;
 
 namespace TurnerSoftware.RobotsExclusionTools.Helpers;