… .appveyor.yml
… .github/workflows/build.yml
… README.md
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/IsValidPathBenchmark.cs
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/PathMatchingBenchmark.cs
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/Resources/RobotsPage-Example.txt
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/RobotsFileParsingBenchmark.cs
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/RobotsPageParsingBenchmark.cs
… benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks/TurnerSoftware.RobotsExclusionTools.Benchmarks.csproj
… build/dotnet-install.ps1
… src/TurnerSoftware.RobotsExclusionTools/Helpers/StreamLineReader.cs
… src/TurnerSoftware.RobotsExclusionTools/IRobotsPageDefinitionParser.cs
… src/TurnerSoftware.RobotsExclusionTools/IsExternalInit.cs
… src/TurnerSoftware.RobotsExclusionTools/PageAccessEntry.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsFile.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsFileAccessRules.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsFileParsedTypes.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsFileParser.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsPageDefinition.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsPageDirectives.cs
… src/TurnerSoftware.RobotsExclusionTools/RobotsPageParser.cs
… src/TurnerSoftware.RobotsExclusionTools/SiteAccessEntry.cs
… src/TurnerSoftware.RobotsExclusionTools/SitemapUrlEntry.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/Extensions.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/ITokenPatternValidator.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/ITokenizer.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/RobotsFileToken.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/RobotsFileTokenReader.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/RobotsPageToken.cs
… src/TurnerSoftware.RobotsExclusionTools/Tokenization/RobotsPageTokenReader.cs