This is a massively breaking change internally and to anything that depended on the internal tokenization/validation system. It is changed in many fundamental ways.

The general public interface is about the same as it was with some more minor binary-breaking changes to methods and constructors.

Closes #42, closes #71 and allows access to the data in #13

Before any changes

|     Method |     Mean |    Error |   StdDev |    Op/s |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|---------:|---------:|--------:|---------:|------:|------:|----------:|
| FromStream | 833.1 us | 14.59 us | 13.65 us | 1,200.3 | 182.6172 |     - |     - |    561 KB |
| FromString | 779.2 us | 15.37 us | 21.04 us | 1,283.3 | 182.6172 |     - |     - |    562 KB |

After the initial tokenization replacement

|     Method |     Mean |   Error |  StdDev |    Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|--------:|--------:|--------:|--------:|------:|------:|----------:|
| FromStream | 178.1 us | 3.33 us | 3.11 us | 5,614.1 | 22.4609 |     - |     - |     69 KB |
| FromString | 170.9 us | 1.40 us | 1.17 us | 5,850.9 | 11.4746 |     - |     - |     35 KB |

After `FromStream` actually processing it via a stream

|     Method |     Mean |   Error |  StdDev |    Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|--------:|--------:|--------:|--------:|------:|------:|----------:|
| FromStream | 262.0 us | 4.68 us | 4.15 us | 3,817.3 | 11.7188 |     - |     - |     36 KB |
| FromString | 174.2 us | 3.28 us | 3.07 us | 5,739.6 | 11.4746 |     - |     - |     35 KB |

After `FromString` sharing the same code path as `FromStream`

|     Method |     Mean |   Error |  StdDev |    Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|--------:|--------:|--------:|--------:|------:|------:|----------:|
| FromStream | 268.0 us | 4.58 us | 4.29 us | 3,731.8 | 11.7188 |     - |     - |     36 KB |
| FromString | 191.6 us | 3.74 us | 4.16 us | 5,218.7 | 11.4746 |     - |     - |     35 KB |

Final Benchmark Results for `RobotsFileParser`

|     Method |      Mean |    Error |   StdDev |     Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
| FromStream | 122.58 us | 2.034 us | 2.498 us |  8,158.1 | 11.7188 |     - |     - |     36 KB |
| FromString |  61.23 us | 0.775 us | 0.725 us | 16,331.7 | 11.4746 |     - |     - |     35 KB |

`FromStream`: 95% faster, 94% fewer allocations
`FromString`: 92% faster, 94% fewer allocations

---

Before any changes

|    Method |     Mean |    Error |   StdDev |     Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
| FromRules | 51.36 us | 0.713 us | 0.632 us | 19,468.8 | 13.4888 |     - |     - |     41 KB |

Initial LINQ-heavy rewrite

|    Method |     Mean |    Error |   StdDev |     Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
| FromRules | 11.00 us | 0.070 us | 0.062 us | 90,949.7 | 3.5248 |     - |     - |     11 KB |

Post-LINQ optimizations and performance improvements

|    Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| FromRules | 4.168 us | 0.0825 us | 0.0883 us | 239,922.8 | 0.8011 |     - |     - |      2 KB |

`FromRules`: 92% faster, 95% fewer allocations