@Turnerj 2022-07-30
Performance improvements for `PathComparisonUtility` tested with .NET 6.

**Before**

|      Method |     Mean |   Error |  StdDev |        Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|--------:|------------:|-------:|------:|------:|----------:|
|      Simple | 265.1 ns | 5.03 ns | 7.83 ns | 3,772,531.2 | 0.0305 |     - |     - |      96 B |
|    Wildcard | 341.4 ns | 2.78 ns | 2.60 ns | 2,929,084.7 | 0.0558 |     - |     - |     176 B |
| MustEndWith | 353.1 ns | 4.08 ns | 3.82 ns | 2,831,954.1 | 0.0458 |     - |     - |     144 B |

**After**

|      Method |      Mean |    Error |   StdDev |         Op/s | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------:|---------:|---------:|-------------:|------:|------:|------:|----------:|
|      Simple |  85.06 ns | 1.528 ns | 1.429 ns | 11,756,811.4 |     - |     - |     - |         - |
|    Wildcard | 128.39 ns | 1.107 ns | 0.924 ns |  7,788,603.9 |     - |     - |     - |         - |
| MustEndWith | 131.07 ns | 1.487 ns | 1.318 ns |  7,629,488.6 |     - |     - |     - |         - |

**Results**
Simple: 3.1x faster
Wildcard: 2.65x faster
MustEndWith: 2.69x faster

All with zero allocations!