@Turnerj 2022-08-01
Current progress has a little bit of a regression - guessing it still is about that hand off to `ProcessLine` etc

|     Method |     Mean |   Error |  StdDev |    Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|--------:|--------:|--------:|--------:|------:|------:|----------:|
| FromStream | 285.5 us | 4.46 us | 3.96 us | 3,502.3 | 11.7188 |     - |     - |     36 KB |
| FromString | 203.1 us | 2.17 us | 1.81 us | 4,924.2 | 11.4746 |     - |     - |     35 KB |