@Turnerj 2022-08-03
After a bunch of back and forth, I realised rather than processing the characters directly in a span, I could process the bytes directly to avoid a bunch of overhead from going back and forth.

|     Method |     Mean |   Error |  StdDev |    Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |---------:|--------:|--------:|--------:|--------:|------:|------:|----------:|
| FromStream | 199.3 us | 3.93 us | 6.00 us | 5,017.7 | 11.7188 |     - |     - |     36 KB |
| FromString | 130.2 us | 1.44 us | 1.20 us | 7,680.2 | 11.2305 |     - |     - |     35 KB |