@Turnerj 2022-08-03
With AVX2 parsing of token values, we can get a significant performance improvement (47% faster) when we have a large search area (eg. `FromString`).

|     Method |      Mean |    Error |   StdDev |     Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
| FromStream | 194.38 us | 3.767 us | 4.031 us |  5,144.5 | 11.7188 |     - |     - |     36 KB |
| FromString |  70.59 us | 1.134 us | 1.061 us | 14,165.6 | 11.3525 |     - |     - |     35 KB |