@Turnerj 2022-08-04
With support of over-reading to fill the vector, we can gain a nice 25% for `FromStream` processing as each line can take advantage of it now.

|     Method |      Mean |    Error |   StdDev |     Op/s |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |----------:|---------:|---------:|---------:|--------:|------:|------:|----------:|
| FromStream | 143.35 us | 1.291 us | 1.144 us |  6,976.0 | 11.7188 |     - |     - |     36 KB |
| FromString |  78.79 us | 0.984 us | 0.920 us | 12,691.7 | 11.3525 |     - |     - |     35 KB |