@Turnerj 2022-08-05
Initial benchmark for `RobotsPageParser`. The file isn't anywhere near as big as the file `RobotsFileParser` needs to process but we really don't expect many rules for individual pages anyway.

|    Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| FromRules | 5.868 us | 0.0678 us | 0.0601 us | 170,406.1 | 2.2659 |     - |     - |      7 KB |