@Turnerj 2022-08-06
More efficient processing of directives by using a mutable structure and dropping all of the LINQ processing.

|    Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------- |---------:|----------:|----------:|----------:|-------:|------:|------:|----------:|
| FromRules | 4.556 us | 0.0497 us | 0.0440 us | 219,504.6 | 0.8011 |     - |     - |      2 KB |

`FromRules` is now ~59% faster and allocate ~82% less